Separated AST nodes into their own files in a ast package. Moved types to this package too
