+++ src/main/kotlin/com/valgolang/frontend/VAlgoLangParserVisitor.kt
@@ -3,6 +3,7 @@ package com.valgolang.frontend
 import antlr.VAlgoLangParser.*
 import antlr.VAlgoLangParserBaseVisitor
 import com.valgolang.errorhandling.semanticerror.incompatibleOperatorTypeError
+import com.valgolang.frontend.ast.*
 import com.valgolang.frontend.datastructures.DataStructureMethod
 import com.valgolang.frontend.datastructures.DataStructureType
 import com.valgolang.frontend.datastructures.ErrorMethod
