+++ src/main/kotlin/com/valgolang/frontend/datastructures/array/ArrayType.kt
@@ -1,9 +1,9 @@
 package com.valgolang.frontend.datastructures.array
 
-import com.valgolang.frontend.BoolType
-import com.valgolang.frontend.NumberType
-import com.valgolang.frontend.Type
-import com.valgolang.frontend.VoidType
+import com.valgolang.frontend.ast.BoolType
+import com.valgolang.frontend.ast.NumberType
+import com.valgolang.frontend.ast.Type
+import com.valgolang.frontend.ast.VoidType
 import com.valgolang.frontend.datastructures.ConstructorMethod
 import com.valgolang.frontend.datastructures.DataStructureMethod
 import com.valgolang.frontend.datastructures.DataStructureType
