+++ src/test/kotlin/com/valgolang/ValidASTTests.kt
@@ -1,6 +1,6 @@
 package com.valgolang
 
-import com.valgolang.frontend.*
+import com.valgolang.frontend.ast.*
 import com.valgolang.frontend.datastructures.stack.StackType
 import junit.framework.TestCase.assertEquals
 import org.junit.jupiter.api.Test
